deleted tests and added pending for ad
